Ballo

+++ src/main/java/com/maliitourist/apigestionregions/apigestionregions/repository/Domaine_activiteRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.maliitourist.apigestionregions.apigestionregions.models.DomaineActivite;
 
-public interface Domaine_activiteRepository extends JpaRepository<DomaineActivite, Integer> {
+public interface Domaine_activiteRepository extends JpaRepository<DomaineActivite, Long> {
     DomaineActivite findByNom(String nom);
 
     DomaineActivite findByCode(String code);
