+++ src/main/java/com/maliitourist/apigestionregions/apigestionregions/repository/PopulationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.maliitourist.apigestionregions.apigestionregions.models.Population;
 
-public interface PopulationRepository extends JpaRepository<Population, Integer> {
-    Population findByCode(int code);
+public interface PopulationRepository extends JpaRepository<Population, Long> {
+    Population findByCode(Long code);
 }
